matrix class

+++ src/loader.cpp
@@ -42,9 +42,9 @@ std::vector<LabeledImage> loadImages(const std::string& imPath, const std::strin
 
     std::println("Loading {} images from {} of size {}x{}", im_count, imPath, im_rows, im_cols);
     std::vector<LabeledImage> images(im_count);
-    uint32_t n = im_rows * im_cols;
-    const uint32_t chunks = im_count < 100 ? 1 : im_count / 100;
-    for (uint32_t im = 0; im < im_count; ++im) {
+    size_t n = im_rows * im_cols;
+    const size_t chunks = im_count < 100 ? 1 : im_count / 100;
+    for (size_t im = 0; im < im_count; ++im) {
         uint8_t d; ls.read(reinterpret_cast<char*>(&d), 1);
         images[im].label = (uint8_t)d;
         for (size_t p = 0; p < n; ++p) {
